Fix #1490: Fix broken sample-data ignestion due to User classes refactored
